+++ docs/src/app/components/pages/components/icon-buttons.jsx
@@ -1,11 +1,10 @@
-const React = require('react');
-const { FontIcon, IconButton, NavigationMenu, Paper } = require('material-ui');
-const ComponentDoc = require('../../component-doc');
-const ActionGrade = require('svg-icons/action/grade');
-const ActionHome = require('svg-icons/action/home');
-const Code = require('icon-buttons-code');
-const CodeExample = require('../../code-example/code-example');
-const CodeBlock = require('../../code-example/code-block');
+import React from 'react';
+import {FontIcon, IconButton, Paper} from 'material-ui';
+import ComponentDoc from '../../component-doc';
+import ActionGrade from 'svg-icons/action/grade';
+import Code from 'icon-buttons-code';
+import CodeExample from '../../code-example/code-example';
+import CodeBlock from '../../code-example/code-block';
 
 export default class IconButtonsPage extends React.Component {
 
@@ -106,6 +105,11 @@ export default class IconButtonsPage extends React.Component {
             header: 'IconButton.onFocus(e)',
             desc: 'Callback function for when the component gains focus.',
           },
+          {
+            name: 'onTouchTap',
+            header: 'IconButton.onTouchTap(e)',
+            desc: 'Callback function for when a touch tap / click event occurs on the button.',
+          },
         ],
       },
     ];
@@ -119,14 +123,14 @@ export default class IconButtonsPage extends React.Component {
         <Paper style = {{marginBottom: '22px'}}>
           <CodeBlock>
           {
-            '//Import statement:\nconst IconButton = require(\'material-ui/lib/icon-button\');\n\n' +
+            '//Import statement:\nimport IconButton from \'material-ui/lib/icon-button\';\n\n' +
             '//See material-ui/lib/index.js for more\n'
           }
           </CodeBlock>
         </Paper>
 
         <CodeExample code={Code}>
-          <IconButton iconClassName="muidocs-icon-custom-github" tooltip="bottom-right"  tooltipPosition = "bottom-right" />
+          <IconButton iconClassName="muidocs-icon-custom-github" tooltip="bottom-right" tooltipPosition = "bottom-right" />
 
           <IconButton iconClassName="muidocs-icon-custom-github" tooltip="bottom-center" tooltipPosition = "bottom-center" />