+++ docs/src/app/components/pages/components/icons.jsx
@@ -1,14 +1,13 @@
-const React = require('react');
-const { ClearFix, FontIcon, Styles, Paper } = require('material-ui');
-const CodeBlock = require('../../code-example/code-block');
-const ComponentDoc = require('../../component-doc');
-const ActionHome = require('svg-icons/action/home');
+import React from 'react';
+import {FontIcon, Styles, Paper} from 'material-ui';
+import CodeBlock from '../../code-example/code-block';
+import ComponentDoc from '../../component-doc';
+import ActionHome from 'svg-icons/action/home';
 
-const { Colors, Typography } = Styles;
-const IconButtonCode = require('icons-code');
-const SvgIconsCode = require('svg-icons-code');
-const CodeExample = require('../../code-example/code-example');
-
+const {Colors, Typography} = Styles;
+import IconButtonCode from 'icons-code';
+import SvgIconsCode from 'svg-icons-code';
+import CodeExample from '../../code-example/code-example';
 
 export default class FontIconPage extends React.Component {
 
@@ -48,14 +47,31 @@ export default class FontIconPage extends React.Component {
     );
 
     let svgIconDesc = (
-      <p style={this.getStyles()}>
-        Alternatively, it is possible to include svg icons using mui.SvgIcon to
-        create a custom svg component. Here we are creating the ActionHome
-        SvgIcon for this docs site, and using it in some seperate component.
-        Custom SvgIcon components can be included as children for other Material
-        UI components that use icons such as <a title="Example of SvgIcon usage"
-        href="#/components/icon-buttons">IconButtons</a>.
-      </p>
+      <div>
+        <p style={this.getStyles()}>
+          Alternatively, it is possible to include svg icons using mui.SvgIcon to
+          create a custom svg component. Here we are creating the ActionHome
+          SvgIcon for this docs site, and using it in some seperate component.
+          Custom SvgIcon components can be included as children for other Material
+          UI components that use icons such as <a title="Example of SvgIcon usage"
+          href="#/components/icon-buttons">IconButtons</a>.
+        </p>
+        <p>
+          For all of the Google's Material Design Icons prebuild SvgIcon components
+          exist. They can be used this way:
+        </p>
+          <CodeBlock>
+            {`
+import SocialNotificationsActive from 'material-ui/lib/svg-icons/social/notifications-active';
+...
+  render: function() {
+    return (
+      <SocialNotificationsActive />
+    );
+  }
+`}
+          </CodeBlock>
+      </div>
     );
 
     let componentInfo = [
@@ -123,7 +139,7 @@ export default class FontIconPage extends React.Component {
           <Paper style = {{marginBottom: '22px'}}>
             <CodeBlock>
             {
-              '//Import statement:\nconst FontIcon = require(\'material-ui/lib/font-icon\');\n\n' +
+              '//Import statement:\nimport FontIcon from \'material-ui/lib/font-icon\';\n\n' +
               '//See material-ui/lib/index.js for more\n'
             }
             </CodeBlock>