+++ docs/src/app/components/pages/components/text-fields.jsx
@@ -1,12 +1,12 @@
-const React = require('react');
-const { ClearFix, Mixins, SelectField, TextField, Styles, Paper } = require('material-ui');
-const ComponentDoc = require('../../component-doc');
-const { Colors } = Styles;
-const { StyleResizable } = Mixins;
-const Code = require('text-fields-code');
-const CodeExample = require('../../code-example/code-example');
-const LinkedStateMixin = require('react-addons-linked-state-mixin');
-const CodeBlock = require('../../code-example/code-block');
+import React from 'react';
+import {ClearFix, Mixins, TextField, Styles, Paper} from 'material-ui';
+import ComponentDoc from '../../component-doc';
+const {Colors} = Styles;
+const {StyleResizable} = Mixins;
+import Code from 'text-fields-code';
+import CodeExample from '../../code-example/code-example';
+import LinkedStateMixin from 'react-addons-linked-state-mixin';
+import CodeBlock from '../../code-example/code-block';
 
 const TextFieldsPage = React.createClass({
 
@@ -21,10 +21,6 @@ const TextFieldsPage = React.createClass({
       propValue: 'Prop Value',
       floatingPropValue: 'Prop Value',
       valueLinkValue: 'Value Link',
-      selectValue: undefined,
-      selectValue2: undefined,
-      selectValueLinkValue: 4,
-      selectValueLinkValue2: 3,
       floatingValueLinkValue: 'Value Link',
     };
   },
@@ -58,6 +54,12 @@ const TextFieldsPage = React.createClass({
         name: 'Props',
         infoArray: [
           {
+            name: 'disabled',
+            type: 'bool',
+            header: 'optional',
+            desc: 'Disables the text field if set to true.',
+          },
+          {
             name: 'defaultValue',
             type: 'string',
             header: 'optional',
@@ -71,9 +73,9 @@ const TextFieldsPage = React.createClass({
           },
           {
             name: 'errorText',
-            type: 'string',
+            type: 'node',
             header: 'optional',
-            desc: 'The error text string to display.',
+            desc: 'The error content to display.',
           },
           {
             name: 'floatingLabelStyle',
@@ -83,9 +85,9 @@ const TextFieldsPage = React.createClass({
           },
           {
             name: 'floatingLabelText',
-            type: 'string',
+            type: 'node',
             header: 'optional',
-            desc: 'The text string to use for the floating label element.',
+            desc: 'The content to use for the floating label element.',
           },
           {
             name: 'fullWidth',
@@ -101,9 +103,9 @@ const TextFieldsPage = React.createClass({
           },
           {
             name: 'hintText',
-            type: 'string',
+            type: 'node',
             header: 'optional',
-            desc: 'The hint text string to display.',
+            desc: 'The hint content to display.',
           },
           {
             name: 'inputStyle',
@@ -119,13 +121,13 @@ const TextFieldsPage = React.createClass({
           },
           {
             name: 'rows',
-            type: 'string',
+            type: 'number',
             header: 'default: 1',
             desc: 'Number of rows to display when multiLine option is set to true.',
           },
           {
             name: 'rowsMax',
-            type: 'string',
+            type: 'number',
             header: 'default: null',
             desc: 'Maximum number of rows to display when multiLine option is set to true.',
           },
@@ -228,20 +230,6 @@ const TextFieldsPage = React.createClass({
     ];
 
     let styles = this.getStyles();
-    let menuItems = [
-      { payload: '1', text: 'Never' },
-      { payload: '2', text: 'Every Night' },
-      { payload: '3', text: 'Weeknights' },
-      { payload: '4', text: 'Weekends' },
-      { payload: '5', text: 'Weekly' },
-    ];
-    let arbitraryArrayMenuItems = [
-      {id:1, name:'Never'},
-      {id:2, name:'Every Night'},
-      {id:3, name:'Weeknights'},
-      {id:4, name:'Weekends'},
-      {id:5, name:'Weekly'},
-    ];
 
     return (
       <ComponentDoc
@@ -252,8 +240,7 @@ const TextFieldsPage = React.createClass({
         <Paper style = {{marginBottom: '22px'}}>
           <CodeBlock>
           {
-            '//Import statement:\nconst TextField = require(\'material-ui/lib/text-field\');\n' +
-            'const SelectField = require(\'material-ui/lib/select-field\');\n\n' +
+            '//Import statement:\nimport TextField from \'material-ui/lib/text-field\';\n\n' +
             '//See material-ui/lib/index.js for more\n'
           }
           </CodeBlock>
@@ -330,31 +317,7 @@ const TextFieldsPage = React.createClass({
                 style={styles.textfield}
                 hintText="Disabled Hint Text"
                 disabled={true}
-                defaultValue="Disabled With Value" /><br/>
-              <SelectField
-                style={styles.textfield}
-                value={this.state.selectValue}
-                onChange={this._handleSelectValueChange.bind(null, 'selectValue')}
-                hintText="Hint Text"
-                menuItems={menuItems} /><br/>
-              <SelectField
-                valueLink={this.linkState('selectValueLinkValue')}
-                floatingLabelText="Float Label Text"
-                valueMember="id"
-                displayMember="name"
-                menuItems={arbitraryArrayMenuItems} /><br/>
-              <SelectField
-                valueLink={this.linkState('selectValueLinkValue2')}
-                floatingLabelText="Float Custom Label Text"
-                floatingLabelStyle={{color: "red"}}
-                valueMember="id"
-                displayMember="name"
-                menuItems={arbitraryArrayMenuItems} /><br/>
-              <SelectField
-                style={styles.textfield}
-                value={this.state.selectValue2}
-                onChange={this._handleSelectValueChange.bind(null, 'selectValue2')}
-                menuItems={arbitraryArrayMenuItems} />
+                defaultValue="Disabled With Value" />
             </div>
             <div style={styles.group}>
               <TextField
@@ -440,12 +403,6 @@ const TextFieldsPage = React.createClass({
     this.setState({
       propValue: e.target.value,
     });
-  },
-
-  _handleSelectValueChange(name, e) {
-    let change = {};
-    change[name] = e.target.value;
-    this.setState(change);
   },
 
   _handleFloatingInputChange(e) {