+++ docs/src/app/components/pages/components/toolbars.jsx
@@ -1,6 +1,6 @@
-const React = require('react');
-const mui = require('material-ui');
-const ComponentDoc = require('../../component-doc');
+import React from 'react';
+import mui from 'material-ui';
+import ComponentDoc from '../../component-doc';
 
 const {
   DropDownIcon,
@@ -13,9 +13,9 @@ const {
   ToolbarTitle,
   Paper,
 } = mui;
-const Code = require('toolbars-code');
-const CodeExample = require('../../code-example/code-example');
-const CodeBlock = require('../../code-example/code-block');
+import Code from 'toolbars-code';
+import CodeExample from '../../code-example/code-example';
+import CodeBlock from '../../code-example/code-block';
 
 export default class ToolbarPage extends React.Component {
 
@@ -92,17 +92,17 @@ export default class ToolbarPage extends React.Component {
     ];
 
     let filterOptions = [
-      { payload: '1', text: 'All Broadcasts' },
-      { payload: '2', text: 'All Voice' },
-      { payload: '3', text: 'All Text' },
-      { payload: '4', text: 'Complete Voice' },
-      { payload: '5', text: 'Complete Text' },
-      { payload: '6', text: 'Active Voice' },
-      { payload: '7', text: 'Active Text' },
+      {payload: '1', text: 'All Broadcasts'},
+      {payload: '2', text: 'All Voice'},
+      {payload: '3', text: 'All Text'},
+      {payload: '4', text: 'Complete Voice'},
+      {payload: '5', text: 'Complete Text'},
+      {payload: '6', text: 'Active Voice'},
+      {payload: '7', text: 'Active Text'},
     ];
     let iconMenuItems = [
-      { payload: '1', text: 'Download' },
-      { payload: '2', text: 'More Info' },
+      {payload: '1', text: 'Download'},
+      {payload: '2', text: 'More Info'},
     ];
 
     return (
@@ -114,10 +114,10 @@ export default class ToolbarPage extends React.Component {
         <Paper style = {{marginBottom: '22px'}}>
           <CodeBlock>
           {
-            '//Import statement:\nconst Toolbar = require(\'material-ui/lib/toolbar/toolbar\');\n' +
-            'const ToolbarGroup = require(\'material-ui/lib/toolbar/toolbar-group\');\n' +
-            'const ToolbarSeparator = require(\'material-ui/lib/toolbar/toolbar-separator\');\n' +
-            'const ToolbarTitle = require(\'material-ui/lib/toolbar/toolbar-title\');\n\n' +
+            '//Import statement:\nimport Toolbar from \'material-ui/lib/toolbar/toolbar\';\n' +
+            'import ToolbarGroup from \'material-ui/lib/toolbar/toolbar-group\';\n' +
+            'import ToolbarSeparator from \'material-ui/lib/toolbar/toolbar-separator\';\n' +
+            'import ToolbarTitle from \'material-ui/lib/toolbar/toolbar-title\';\n\n' +
             '//See material-ui/lib/index.js for more\n'
           }
           </CodeBlock>