+++ src/date-picker/date-picker.jsx
@@ -21,7 +21,7 @@ const DatePicker = React.createClass({
     muiTheme: React.PropTypes.object,
   },
 
-  getChildContext () {
+  getChildContext() {
     return {
       muiTheme: this.state.muiTheme,
     };
@@ -48,6 +48,7 @@ const DatePicker = React.createClass({
     showYearSelector: React.PropTypes.bool,
     style: React.PropTypes.object,
     textFieldStyle: React.PropTypes.object,
+    value: React.PropTypes.object,
   },
 
   windowListeners: {
@@ -124,7 +125,7 @@ const DatePicker = React.createClass({
           initialDate={this.state.dialogDate}
           onAccept={this._handleDialogAccept}
           onShow={onShow}
-          onDismiss={this._handleDialogDismiss}
+          onDismiss={onDismiss}
           minDate={minDate}
           maxDate={maxDate}
           autoOk={autoOk}
@@ -171,10 +172,6 @@ const DatePicker = React.createClass({
     }
     if (this.props.onChange) this.props.onChange(null, d);
     if (this.props.valueLink) this.props.valueLink.requestChange(d);
-  },
-
-  _handleDialogDismiss() {
-    if (this.props.onDismiss) this.props.onDismiss();
   },
 
   _handleInputFocus(e) {