+++ src/table/table.jsx
@@ -27,6 +27,10 @@ const Table = React.createClass({
     onRowClick: React.PropTypes.func,
     selectable: React.PropTypes.bool,
     style: React.PropTypes.object,
+    wrapperStyle: React.PropTypes.object,
+    headerStyle: React.PropTypes.object,
+    bodyStyle: React.PropTypes.object,
+    footerStyle: React.PropTypes.object,
   },
 
   getDefaultProps() {
@@ -45,13 +49,13 @@ const Table = React.createClass({
     muiTheme: React.PropTypes.object,
   },
 
-  getChildContext () {
+  getChildContext() {
     return {
       muiTheme: this.state.muiTheme,
     };
   },
 
-  getInitialState () {
+  getInitialState() {
     return {
       muiTheme: this.context.muiTheme ? this.context.muiTheme : ThemeManager.getMuiTheme(DefaultRawTheme),
       allRowsSelected: this.props.allRowsSelected,
@@ -60,7 +64,7 @@ const Table = React.createClass({
 
   //to update theme inside state whenever a new theme is passed down
   //from the parent / owner using context
-  componentWillReceiveProps (nextProps, nextContext) {
+  componentWillReceiveProps(nextProps, nextContext) {
     let newMuiTheme = nextContext.muiTheme ? nextContext.muiTheme : this.state.muiTheme;
     this.setState({muiTheme: newMuiTheme});
   },
@@ -100,6 +104,10 @@ const Table = React.createClass({
       fixedFooter,
       fixedHeader,
       style,
+      wrapperStyle,
+      headerStyle,
+      bodyStyle,
+      footerStyle,
       ...other,
     } = this.props;
     let classes = 'mui-table';
@@ -130,7 +138,7 @@ const Table = React.createClass({
     let inlineHeader, inlineFooter;
     if (fixedHeader) {
       headerTable = (
-        <div className="mui-header-table">
+        <div className="mui-header-table" style={this.prepareStyles(headerStyle)}>
           <table className={className} style={mergedTableStyle}>
             {tHead}
           </table>
@@ -143,7 +151,7 @@ const Table = React.createClass({
     if (tFoot !== undefined) {
       if (fixedFooter) {
         footerTable = (
-          <div className="mui-footer-table">
+          <div className="mui-footer-table" style={this.prepareStyles(footerStyle)}>
             <table className={className} style={mergedTableStyle}>
               {tFoot}
             </table>
@@ -156,9 +164,9 @@ const Table = React.createClass({
     }
 
     return (
-      <div className="mui-table-wrapper" style={this.prepareStyles(styles.tableWrapper)}>
+      <div className="mui-table-wrapper" style={this.prepareStyles(styles.tableWrapper, wrapperStyle)}>
         {headerTable}
-        <div className="mui-body-table" style={this.prepareStyles(styles.bodyTable)} ref="tableDiv">
+        <div className="mui-body-table" style={this.prepareStyles(styles.bodyTable, bodyStyle)} ref="tableDiv">
           <table className={classes} style={mergedTableStyle} ref="tableBody">
             {inlineHeader}
             {inlineFooter}
@@ -232,7 +240,7 @@ const Table = React.createClass({
   },
 
   _onRowSelection(selectedRows) {
-    if (this.state.allRowsSelected) this.setState({ allRowsSelected: false });
+    if (this.state.allRowsSelected) this.setState({allRowsSelected: false});
     if (this.props.onRowSelection) this.props.onRowSelection(selectedRows);
   },
 
@@ -248,7 +256,7 @@ const Table = React.createClass({
         this.props.onRowSelection('none');
       }
     }
-  
+
     this.setState({allRowsSelected: !this.state.allRowsSelected});
   },