+++ src/toggle.jsx
@@ -20,6 +20,12 @@ const Toggle = React.createClass({
     onToggle: React.PropTypes.func,
     toggled: React.PropTypes.bool,
     defaultToggled: React.PropTypes.bool,
+    thumbStyle: React.PropTypes.object,
+    iconStyle: React.PropTypes.object,
+    labelPosition: React.PropTypes.oneOf(['left', 'right']),
+    trackStyle: React.PropTypes.object,
+    disabled: React.PropTypes.bool,
+    rippleStyle: React.PropTypes.object,
   },
 
   //for passing default theme context to children
@@ -27,7 +33,7 @@ const Toggle = React.createClass({
     muiTheme: React.PropTypes.object,
   },
 
-  getChildContext () {
+  getChildContext() {
     return {
       muiTheme: this.state.muiTheme,
     };
@@ -46,7 +52,7 @@ const Toggle = React.createClass({
 
   //to update theme inside state whenever a new theme is passed down
   //from the parent / owner using context
-  componentWillReceiveProps (nextProps, nextContext) {
+  componentWillReceiveProps(nextProps, nextContext) {
     let newMuiTheme = nextContext.muiTheme ? nextContext.muiTheme : this.state.muiTheme;
     this.setState({muiTheme: newMuiTheme});
   },
@@ -67,11 +73,11 @@ const Toggle = React.createClass({
         width: toggleTrackWidth,
       },
       track: {
-          transition: Transitions.easeOut(),
-          width: '100%',
-          height: 14,
-          borderRadius: 30,
-          backgroundColor: this.getTheme().trackOffColor,
+        transition: Transitions.easeOut(),
+        width: '100%',
+        height: 14,
+        borderRadius: 30,
+        backgroundColor: this.getTheme().trackOffColor,
       },
       thumb: {
         transition: Transitions.easeOut(),
@@ -160,8 +166,8 @@ const Toggle = React.createClass({
     );
 
     let enhancedSwitchProps = {
-      ref: "enhancedSwitch",
-      inputType: "checkbox",
+      ref: 'enhancedSwitch',
+      inputType: 'checkbox',
       switchElement: toggleElement,
       rippleStyle: customRippleStyle,
       rippleColor: rippleColor,
@@ -173,7 +179,7 @@ const Toggle = React.createClass({
       onSwitch: this._handleToggle,
       onParentShouldUpdate: this._handleStateChange,
       defaultSwitched: this.props.defaultToggled,
-      labelPosition: (this.props.labelPosition) ? this.props.labelPosition : "left",
+      labelPosition: (this.props.labelPosition) ? this.props.labelPosition : 'left',
     };
 
     if (this.props.hasOwnProperty('toggled')) enhancedSwitchProps.checked = this.props.toggled;