+++ CHANGELOG.md
@@ -1,3 +1,37 @@
+## 0.13.4
+###### _Nov 24, 2015_
+
+##### General
+- Introduced SelectableEnhance HOC to wrap List with valueLink (#1976)
+- Added color prop to LinearProgress and RefreshIndicator (#2206)
+- [AutoComplete] new component! (#2187) (thanks @yongxu)
+- [Table] added wrapperStyle prop to override table wrapper's styles (#2238)
+- Updated SVG icons (#2240)
+- [Table] added props for headerStyle, bodyStyle and footerStyle (#2246)
+
+##### Component Fixes / Enhancements
+- Fixed double ripple due to compat mouse down (#2216)
+- [RenderToLayer] iframe support for clickaway (#2210)
+- [TextField] Fixed floating label element not allowing focus (#2228)
+- [SelectField] onFocus and onBlur handlers passed to underlying TextField component (#2102)
+
+## 0.13.3
+###### _Nov 17, 2015_
+
+##### General
+- [Snackbar] add bodyStyle prop to style child div (#2104)
+- [DatePicker] add container prop to display DatePicker in-line or inside Dialog (#2120 and #2153)
+- [AppBar] add relative positioning for z-index to take effect (#1478)
+- [AppBar] add onTitleTouchTap prop to AppBar (#2125)
+- [Popover] new component! (#2043) (thanks @chrismcv)
+- Split [SelectField] and [TextField] doc pages (#2161)
+
+##### Component Fixes / Enhancements
+- [SelectField] onChange triggered consistently when using value prop (#1610)
+- [Dialog] fix page scrolling behind dialog after resizing (#1946)
+- [DatePicker] fix calendar height (#2141)
+- [TimePicker] allow to set time to null (#2108)
+
 ## 0.13.2
 ###### _Nov 9, 2015_